ocr-backend/requirements.txt
fastapi
uvicorn
pytesseract
opencv-python
pillow
python-multipart